requirements.txt
firebase-admin
opencv-python
numpy
imutils
openai==1.3.5
jdeskew
PyMuPDF
flask-cors
pillow
pdf2image
Flask==2.2.2
werkzeug==2.3.0
gunicorn
functions_framework
matplotlib
plotly
scipy
scikit-learn
js2py
pandas
nodejs-bin
tiktoken
google-cloud-aiplatform